requirements.txt
asyncpool==1.0
psycopg2==2.8.6
PyYAML==5.4.1
